+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## main(v1.0.1)
 
 - feat: |UI| 增加极简模式主页, 可在 `外观` 中切换
+- fix: 修复 oauth2 登录时，default role 不生效的问题
 
 ## v1.0.0